refactor: Extract into `TaxonDecorator::HeadlineProtonym`
